열 일곱 번째 커밋
routes/events.py UPDATE, DELETE 라우트 수정

+++ routes/events.py
@@ -43,20 +43,29 @@ async def create_event(body:Event, user: str = Depends(authenticate)) -> dict:
 # UPDATE 라우트
 @event_router.put("/{id}", response_model=Event)
 async def update_event(id: PydanticObjectId, body:EventUpdate, user: str=Depends(authenticate)) -> Event:
-    print("PUT 호출되었습니다.")
     updated_event=await event_database.update(id, body)
+    event=await event_database.get(id)
     if not updated_event:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Event with supplied ID does not exist."
         )
-
+    if event.creator!=user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed."
+        )
     return updated_event
-
 
 # 데이터베이스에 있는 단일 이벤트 삭제
 @event_router.delete("/{id}")
 async def delete_event(id:PydanticObjectId, user:str=Depends(authenticate)) -> dict:
+    event=await event_database.get(id)
+    if event.creator!=user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Operation not allowed."
+        )
     event=await event_database.delete(id)
     if not event:
         raise HTTPException(
